requirements.txt
asgiref==3.5.2
distlib==0.3.4
Django==3.2.14
django-braces==1.15.0
django-crispy-forms==1.14.0
django-filter==22.1
django-gravatar2==1.4.4
django-model-utils==4.2.0
django-star-ratings==0.9.2
django-utils-six==2.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
filelock==3.7.1
Markdown==3.4.1
pbr==5.9.0
Pillow==9.1.1
platformdirs==2.5.2
python-decouple==3.6
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
stevedore==3.5.0
swapper==1.3.0
virtualenv==20.15.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
